requirements.txt
numpy==1.21.5
pandas==1.3.5
matplotlib
streamlit==1.1.0
opencv-python-headless
tensorflow-cpu